Implement P2.4 bounded local concentration-tempering and update M2 build plan
- Added P2.4 bounded local concentration-tempering snapshot to segment_5B.build_plan.md, detailing trigger execution, implementation posture, candidate grid outcomes, and closure decision.
- Updated definition of done to reflect closure of P2.4 lane.
- Expanded segment_5B.impl_actual.md with execution updates for P2.4, including candidate protocol execution and canonical restore decision.
- Created new platform.M2.build_plan.md for M2 deep planning, detailing purpose, authority inputs, scope boundaries, deliverables, execution gates, and work breakdown.
- Updated platform.build_plan.md to reflect M2 status as IN_PROGRESS and added detailed sub-phase progress.
- Expanded platform.impl_actual.md with entries documenting M2 planning decisions, execution kickoffs, and closure verdicts.
- Created minimal Terraform stack roots for core, data_ml, ops, runtime, and streaming to satisfy M2.A backend/init readiness checks.
- Updated runner.py to set explicit return_dtype on tzid_operational mapping to tighten deterministic dtype behavior.

+++ packages/engine/src/engine/layers/l2/seg_5B/s4_arrival_events/runner.py
@@ -1584,7 +1584,9 @@ def run_s4(config: EngineConfig, run_id: Optional[str] = None) -> S4Result:
         alias_layout = alias_layout_policy
         endianness = str(alias_layout.get("endianness") or "little")
         edge_catalogue_df = edge_catalogue_df.with_columns(
-            pl.col("tzid_operational").map_elements(lambda v: tzid_to_idx.get(str(v), -1)).alias(
+            pl.col("tzid_operational").map_elements(
+                lambda v: tzid_to_idx.get(str(v), -1), return_dtype=pl.Int32
+            ).alias(
                 "tzid_operational_idx"
             )
         )
